split function naming and api route + added version number
